Added group test; restuctured 'test' file tree
